requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
Authlib==1.3.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
defusedxml==0.7.1
Flask==3.1.0
grpcio==1.68.1
grpcio-health-checking==1.68.1
grpcio-tools==1.68.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jira==3.8.0
kombu==5.4.2
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.48
protobuf==5.29.2
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
validators==0.34.0
vine==5.1.0
wcwidth==0.2.13
weaviate-client==4.10.2
Werkzeug==3.1.3
